accedentally had integer as obj data type in database, switch to real

+++ database/schema.sql
@@ -23,7 +23,7 @@ CREATE TABLE all_solutions (
     problem_instance_name TEXT NOT NULL,
     submission_time DATETIME NOT NULL,
     validation_end_time DATETIME NOT NULL,
-    objective_value INTEGER DEFAULT NULL,
+    objective_value REAL DEFAULT NULL,
     reward_accumulated INTEGER DEFAULT 0,
     accepted BOOLEAN DEFAULT NULL,   -- whether the solution was accepted as current best one or not (NULL if not yet evaluated)
     sol_file_path TEXT,   -- location of the solution file (path would only exist during solution validation phase)
@@ -37,7 +37,7 @@ CREATE TABLE active_solutions_submissions_validations (
     problem_instance_name TEXT,
     agent_validated_id TEXT NOT NULL,   -- id of the agent that validated the solution
     validation_response BOOLEAN NOT NULL,   -- whether the agent accepted (1) or rejected (0) the solution
-    objective_value INTEGER NOT NULL,   -- objective value calulated by the agent
+    objective_value REAL NOT NULL,   -- objective value calulated by the agent
     reward INTEGER NOT NULL,   -- reward that should be given to the agent who validated by the agent
     PRIMARY KEY (solution_submission_id, agent_validated_id),   -- each agent can only validate each solution submission once
     FOREIGN KEY (solution_submission_id) REFERENCES all_solutions (id),
